--- workstation/src/main/java/io/xj/gui/types/CompilationState.java
@@ -1,9 +1,0 @@
-package io.xj.gui.types;
-
-public enum CompilationState {
-  Standby,
-  Compiling,
-  Done,
-  Failed,
-  Cancelled,
-}
